GTC-3144 Create carbon emissions titiler endpoint
This is using the colors/symbology from the existing GFW raster tile
cache.

It includes one filter which is tree cover density, using
umd_tree_cover_density_2010/v1.6. This actually needs to be the 2000
version, I will change that before checkin, after I generate the proper
COGs.

I added an option in reader.py to specify the default_band, since it's
possible that the default band is not called default.tif.

All rendering is true_color (not encoded), so there is no render_type
parameter.