GTC-3158 Carbox flux titiler endpoint
This is an attempt where we separate out the tree cover density, tree
cover gain, and mangrove stock filters, so we need only one base carbon
raster. We don't need to include the pre-2000 tree plantations as a
negative filter - those areas have already been masked-out in the input
data set.

One thing I had to deal with is that the COG reader can return an
NonExistentTile exception if the entire tile area is no-data. This
especially happens for the mangroves filter. So, I added code to check
if the tile exists, and set the data to None if it doesn't exist. Then
the masking code deals with the value of None meaning all no-data. For
our conditions, which are all OR'ed, this means just skipping a filter
if its tile is all no-data.

We may eventually want to make the colormap processing code faster if
possible, since this colormap is quite large.

Tested locally.